DATABASE INIT - Рефактор под бд - Очень много проблем
Реализовано:
- Домашняя страница
- Страница добавления фильма

TODO:
- Рефактор страницы фильм детаилс
- Рефактор страницы филмс из жанр

+++ app/src/main/java/com/leonardower/mymovie/ui/components/common/RatingButton.kt
@@ -17,7 +17,7 @@ fun RatingButton(
     modifier: Modifier = Modifier,
     backgroundColor: Color = GrayButtonColor,
     isRated: Boolean = false,
-    rating: Int? = null,
+    rating: Float? = null,
     enabled: Boolean = true
 ) {
     val buttonText = if (isRated && rating != null) {
@@ -57,7 +57,7 @@ private fun Preview() {
         RatingButton(
             onClick = {},
             isRated = true,
-            rating = 8
+            rating = 8f
         )
     }
 }
